requirements.txt
transformers>=4.35
autoawq

# others:
chardet